phase_3_sense_layer/requirements.txt
google-generativeai
pydantic
pytest
python-dotenv
